Add support for internal authentication
As part of spinning out the attachment API from matchmaker we need services to be able to authenticate internally without having user authentication. This is because the matchmaker indexer needs access to the attachment api which needs access to the identity service.

This change implements a second authentication provider for all routes. Note this is a major version bump because of required environment variables being introduced.